+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Version 1.6.0-UNRELEASED
 * Fix directory creation with SFTP
 * Fix submodule handling
 * Fix/Add support for nested branch names by allowing `/` in scope names
+* Fix error message for unrecognised options
 * Pass insecure-flag to submodules
 * Pass ssh-keys to submodules if used
 * Add support for config `disable-epsv` and `no-commit`